Основной сервис отправляет запросы в статистику. Поправил несколько тестов.

+++ main-service/src/main/java/ru/practicum/dto/EventShortDto.java
@@ -32,8 +32,8 @@ public class EventShortDto {
     private String title;
 
     @PositiveOrZero
-    private Integer confirmedRequests = 0;
+    private long confirmedRequests = 0;
 
     @PositiveOrZero
-    private Integer views = 0;
+    private long views = 0;
 }
